pyproject.toml
[project]
name = "grounded-multimodal-analyst"
version = "0.1.0"
description = "Grounded QA over mixed documents and tables with citations."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.2",
  "openpyxl>=3.1.5",
  "pypdf>=4.3.1",
  "python-docx>=1.1.2",
  "scikit-learn>=1.5.1",
  "numpy>=1.26.4",
  "python-dotenv>=1.0.1",
  "streamlit>=1.39.0"
]

[project.scripts]
grounded-analyst = "grounded_analyst.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
